Spend the string-COMDAT oracle: 30 functions, and no code moved
An MSVC `??_C@...` name is a hash of the string it labels, so when our object
and retail's disagree on that name the TEXT differs -- and retail's object
carries the text. That makes retail the oracle for what our source should say,
which is the difference between reading a fix off and guessing it.

Turning on `functionRelocDiffs=name_check` made those disagreements visible for
the first time: 111 pairs across 98 units, decoded in
<decomp-bench>/archive/runs/namecheck-flip-three-projects-20260812/data/dc3_string_comdat_oracle.txt.
This is the pure-text tranche of them.

  matched_code (name_check)   40.3618% -> 40.4567%   (+10,792 bytes)
  30 functions newly complete, 0 newly incomplete
  matched_code (none)         43.727554% both sides, byte-identical

That last line is the proof, not a footnote: none of this moved a single
instruction. These were always relocation-target differences, and the shipped
ruler could not see any of them.

Most are MILO_ASSERT expressions, which the macro stringizes -- so the assert's
TEXT is its source text, and matching retail means spelling the expression the
way retail spelled it: `(0.0f <= gainReq) && (gainReq <= 1.0f)` rather than the
same thing unparenthesised, `buf[buf_sz-1]` rather than `buf[buf_sz - 1]`,
`mDataPoint == NULL` rather than `!mDataPoint`.

Worth naming individually:

  gBinStream -> gStream in DataFile.cpp. Deferred from the last wave precisely
  because it changes an assert's text; now the assert text is the evidence FOR
  it -- retail's reads `!gStream->Fail()` -- and both witnesses agree.

  Named constants we had inlined, restored: kMaxTasksOneTime for 0x40,
  MIN_SONG_VERSION / MAX_SONG_VERSION for 0 / 99, RndMat::kColorModNum for
  BaseMaterial's spelling of the same inherited constant.

  The MIN/MAX_SONG_VERSION pair needed more than a respelling: they were
  `#define`s, and MILO_ASSERT_RANGE has no `#` in its own body, so its min/max
  arguments get macro-prescanned and the literals were baked into the text
  before the stringize ever happened. As enumerators there is nothing to
  expand, so the identifiers survive. Same values, and worth knowing about for
  the next one.

  SYNC_PROP keys: RndAnimFilter synced mType under `type` where retail uses
  `anim_type`, UIGuide the same under `direction`. These are the reflection
  names data files address, so ours were wrong for content as well as bytes.

  Magnu::EditToken said "Magnu add token only supported on PC" -- copy-pasted
  from AddToken directly above it, which correctly says "add".

  StoreEnumeration::IsSuccess had `if (mHandle != 0) MILO_ASSERT(false, ...)`,
  which is the decompiler's unwrapping of `MILO_ASSERT(!mHandle, ...)`. Same
  condition, same effect, and retail's text confirms the original shape.

  curl's override_userpass reported `_netrc`, because this build forces WIN32
  and setup.h then defines DOT_CHAR as "_". Retail shipped ".netrc", so the
  literal is spelled out at that one message rather than changing the macro.

Reverted after measuring: inlining `Abs(1.0f - Length(inVector))` into
RndAmbientOcclusion::BuildSHCoeff's assert, the way retail spells it, dropped
that function from 100% to 92.5% at the `none` ruler. The hoisted `diff` local
is load-bearing for codegen, so the byte-exact spelling wins over the
text-exact one and the assert text stays on the open list.

Not applied, and each is a real finding rather than a skip:

  SpotlightDrawer::DrawShadow asserts shadowSpot->TargetShadow() where retail
  asserts GetCastShadow(). No such method exists anywhere in this tree, and
  TargetShadow() returns a separately tracked member, so this is a missing or
  renamed method on Spotlight, not a spelling.

  HolmesClient's __FUNCTION__ reads `anonymous-namespace'::WaitForAnyResponse
  against retail's Holmes::WaitForAnyResponse: retail had that function in
  namespace Holmes and we have it anonymous. A namespace move, not a string.

  XMARKETPLACE_MAX_OFFERIDS does not exist in this tree (no xmarketplace XDK
  header at all), so StorePurchaser keeps the literal 6.

  Trie::check_index lives in trie.h, and retail's `0<= n && n < MAX_NODES`
  implies a SIGNED parameter where ours is `unsigned int index` -- porting the
  lower bound onto an unsigned type would be a no-op. That is a signature
  question, not a text one.

  Voice::Stop asserts mEnvelopeParams where retail asserts mPoolVoice.egParams.
  Our Voice holds a flat member at the matching offset instead of an embedded
  PoolVoice, so matching the text means changing the class layout.

+++ src/system/hamobj/HamNavProvider.cpp
@@ -373,7 +373,7 @@ DataNode HamNavProvider::OnSetFormatArgs(const DataArray *a) {
     } else {
         index = FindLabel(evalNode.ForceSym());
     }
-    MILO_ASSERT(index >= 0 && (unsigned int)index < (unsigned int)mNavItems.size(), 0x16d);
+    MILO_ASSERT(index >= 0 && index < mNavItems.size(), 0x16d);
 
     if (mNavItems[index].mFormatArgs) {
         mNavItems[index].mFormatArgs->Release();
